일반콜 호출, 호출정보 조회, 확정/미확정 호출 취소 API 적용 및 UI 적용

+++ lib/data/model/work/reservation_info_response.dart
@@ -44,10 +44,10 @@ class Reservation {
   String? reqEndPlaceNm;
   int? drvPaymPrice;
   int? drvDistance;
-  int? gpsStartLat;
-  int? gpsStartLong;
-  int? gpsEndLat;
-  int? gpsEndLong;
+  double? gpsStartLat;
+  double? gpsStartLong;
+  double? gpsEndLat;
+  double? gpsEndLong;
 
   Reservation({
     required this.drvReqSq,
